Fix snapshot test
