+++ Sources/PlaybookUI/PlaybookGallery.swift
@@ -11,8 +11,6 @@ public struct PlaybookGallery: View {
     private var galleryState = GalleryState()
     @StateObject
     private var imageLoader = ImageLoader()
-    @FocusState
-    private var isFocused
 
     public init(
         title: String? = nil,
@@ -23,89 +21,9 @@ public struct PlaybookGallery: View {
     }
 
     public var body: some View {
-        NavigationView {
-            List {
-                Group {
-                    Spacer.fixed(length: 16)
-                    SearchBar(text: $searchState.query)
-                        .focused($isFocused)
-
-                    Counter(
-                        count: searchState.result.count,
-                        total: searchState.result.total
-                    )
-                    .onDisappear {
-                        isFocused = false
-                    }
-
-                    if searchState.result.kinds.isEmpty {
-                        UnavailableView(
-                            symbol: .magnifyingglass,
-                            description: "No Result for \"\(searchState.query)\""
-                        )
-                    }
-                    else {
-                        ForEach(searchState.result.kinds, id: \.kind) { data in
-                            GalleryKindRow(data: data) { selected in
-                                galleryState.selected = SelectData(
-                                    kind: selected.kind,
-                                    scenario: selected.scenario
-                                )
-                            }
-                        }
-                    }
-
-                    Spacer.fixed(length: 24)
-                }
-                .listRowSpacing(.zero)
-                .listRowInsets(EdgeInsets())
-                .listRowSeparator(.hidden)
-                .listRowBackground(Color.clear)
-                .transition(.identity)
-            }
-            .listStyle(.plain)
-            .environment(\.defaultMinListRowHeight, 0)
-            .navigationTitleIfPresent(title)
-            .background {
-                Color(.background)
-                    .ignoresSafeArea()
-            }
-            .ignoresSafeArea(.keyboard)
-            .sheet(item: $galleryState.selected) { data in
-                GalleryDetail(data: data)
-            }
-            .toolbar {
-                ToolbarItem(placement: .topBarTrailing) {
-                    HStack {
-                        Menu {
-                            Button("Clear Thumbnail Cache") {
-                                galleryState.clearImageCache()
-                            }
-                        } label: {
-                            Image(symbol: .ellipsisCircle)
-                                .imageStyle(font: .subheadline)
-                        }
-
-                        ColorSchemePicker(colorScheme: $galleryState.colorScheme)
-                    }
-                }
-            }
-        }
-        .navigationViewStyle(.stack)
-        .preferredColorScheme(galleryState.colorScheme)
-        .environmentObject(imageLoader)
-    }
-}
-
-@available(iOS 14.0, *)
-private extension View {
-    @ViewBuilder
-    func navigationTitleIfPresent<S: StringProtocol>(_ title: S?) -> some View {
-        if let title {
-            navigationTitle(title)
-        }
-        else {
-            self
-        }
+        PlaybookGalleryContent(title: title)
+            .environmentObject(searchState)
+            .environmentObject(galleryState)
+            .environmentObject(imageLoader)
     }
 }
