+++ Tests/AllScenarios.swift
@@ -1,5 +1,6 @@
 import PlaybookUI
 
+@available(iOS 15.0, *)
 struct AllScenarios: ScenarioProvider {
     static func addScenarios(into playbook: Playbook) {
         playbook
